Revert "feat: class 1 & 2"
