--- src/js/component/Footer.jsx
@@ -1,22 +1,0 @@
-// 1. importar react
-import React from "react";
-
-
-// 2. Crear componente
-const Footer = () => {
-  // 3. Codihgo JS
-
-  // 4. return 1 solo elemento HTML
-  return (
-    <p>
-      <hr />
-				Made by{" "}
-				<a href="http://www.4geeksacademy.com">4Geeks Academy</a>, with
-				love!
-		</p>
-  )
-}
-
-
-// 5. exportar 
-export default Footer;
